BOTS/09.Image_Generator/requirements.txt
huggingface_hub
python-dotenv
Pillow
fastapi
uvicorn[standard]
pydantic
pydantic-settings
streamlit
requests